Create Arduino Java wrapper and change serial communication to use Json objects

+++ src/ouscr/mercury/networking/Frame.java
@@ -21,8 +21,8 @@ public class Frame implements Serializable
     }
 
     public static class RobotInstruction implements Serializable {
-        public float leftMotor;
-        public float rightMotor;
+        public int leftMotor;
+        public int rightMotor;
     }
 
     public FrameType type = FrameType.UNKNOWN;
